Initial Commit - A group of files designed to copy changed sectors in memory.
copyblocks.c is the driver file, which reads the piped input file, giving a completed tree to
copychanges.c, which iterates through the tree and copies the specified sectors to an output file.
copychanges.c consists solely of two methods, copy_sector_range and copychanges,the latter of which calls the former.
rbtree.c, rangetree.c, and the two *.h files create the red/black range tree that stores the
identified hard disk sectors.