Pretty much a complete rewrite
Using ioctl for obtaining blocksize, size, rotational, etc., instead of /sys/fs.  The "gurus"
recommend the /sys/fs approach, but it doesn't work well for us in this case.  Changed
case of a string literal (a const string, not a define), and added underscores to scratch
variables.

Changed the stat() call to a lstat() call in order to cause symlinks to block devices to fail