Adding new POC for OO protobuf communication
Added Proof of Concept for an OO interface to the Linux socket API that
greatly simplifies the process of sending and receiving Google Protocol
Buffers via sockets.

The ProtoBuffer class can utilize UNIX and TCP/IP sockets.  It has the ability
to listen for new connections, connect to remote servers, poll() to prevent
blocking, sending and receiving messages, etc.