Helper script check_guard as requested by Nick
check_guard checking.  Checks for #ifndef/#define lines as first "meaningful" lines of a
header file.  Noticed that it does throw false positives on a few files in the repository
because they use C-style comment lines.

This script will accept the name of a single header file as an argument, and check it.  It will also
accept multiple header file names and check them all.  It will also accept a directory name as an
argument, and recursively check all header files underneath that directory.